minor fixes plus sortingFunctions fully completed
